+++ server/views/index.hbs
@@ -22,9 +22,7 @@
         <title>{{title}}</title>
         <meta name="viewport" content="initial-scale=1, width=device-width" />
         <link rel="stylesheet" href="/static/robotto_font/Robotto.css" />
-        <link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.1/dist/leaflet.css" />
         <link rel="icon" href="/static/favicon.png" type="image/png">
-        <script src="https://api-maps.yandex.ru/2.1/?427e52d5-1d06-4cca-a71b-114ee4358200&lang=ru_RU" type="text/javascript">
     </script>
     </head>
