+++ source/client/src/components/AppForm/Devs/Devs.tsx
@@ -13,9 +13,7 @@ import LogoutRoundedIcon from "@mui/icons-material/LogoutRounded";
 import ModeEditRoundedIcon from "@mui/icons-material/ModeEditRounded";
 
 import DirectionsIcon from '@mui/icons-material/Directions';
-
-
-
+import { MenuDevs } from "./MenuDevs";
 interface IProps {
 }
 
@@ -24,62 +22,7 @@ interface IProps {
 @observer
 export class Devs extends React.Component<IProps> {
   constructor(props: any) {
-    super(props);
-  }
-
-  async editDevice(a:any) { /////////////////////////////////////////Кнопка редактирования
-    APP_STORAGE.devs.setMenu_devs(a);
-    let DevGr: any;
-    DevGr = APP_STORAGE.devs_groups.getDevsGroups();
-    return this.PassValueEditForm(DevGr); 
-  }
-
-
-
-  PassValueEditForm(dgrs: TDevsGroup[]) {
-    var dev = APP_STORAGE.devs;
-    var devGr = APP_STORAGE.devs_groups;
-    var parent: React.ReactNode[] = new Array();
-    for (var ii in dgrs) {
-      var dgr: TDevsGroup = dgrs[ii];
-      var gr: TDGroup = dgr.group;
-      var gr_childs = dgr.childs;
-      var gr_devs = dgr.devs;
-      var childs: React.ReactNode[] = new Array();
-      if (gr_childs.length > 0) childs = this.PassValueEditForm(gr_childs);
-
-      parent.push(childs);
-      for (var key in gr_devs) {
-        if (
-          "_dev_id_key_" + gr_devs[key].id === APP_STORAGE.devs.getIdChild()
-        ) {
-          dev.setNumber(String(gr_devs[key].number));
-          devGr.setParentId('key-09');
-          dev.setName(String(gr_devs[key].name));
-          dev.setLongitude(String(gr_devs[key].longitude));
-          dev.setLatitude(String(gr_devs[key].latitude));
-          dev.setInfo(String(gr_devs[key].info));
-          dev.setChangeSensors(gr_devs[key].sensors.s);
-          dev.setId(String(gr_devs[key].id));
-          dev.setGroupDevId(String(gr_devs[key].group_dev_id));
-          dev.setCheckboxEd(gr_devs[key].deleted);
-            
-
-      if(dev.getMenu_devs() === '1')
-          {dev.setOpenModalChange(true);
-            devGr.setOpen_menu(false);
-          }
-      if(dev.getMenu_devs() === '2'){
-        devGr.setOpenModalMoveDevsGr(true);
-        devGr.setOpen_menu(false);
-          }
-          
-        }
-      }
-    }
-    return parent;
-  }
-
+    super(props);}
 
 
   drawDevs(dgrs: TDevsGroup[]): React.ReactNode[] {
@@ -97,6 +40,7 @@ export class Devs extends React.Component<IProps> {
 
             if ("_dev_id_key_" + gr_devs[key].id === dev.getIdChild()) {
               APP_STORAGE.sensors.setNumber(gr_devs[key].number);
+
                   if (gr_devs[key].deleted === true) {
 
                                     devs.push(
@@ -117,44 +61,7 @@ export class Devs extends React.Component<IProps> {
                                     </Typography>
                                     
                                     {APP_STORAGE.getRoleWrite() === 2 &&  APP_STORAGE.getRoleRead() === 1 && (
-                                    <div>
-                                    <IconButton
-                                    onClick={() => {devGr.setOpen_menu(true)}}
-                                    id="long-button-menu"
-                                    aria-label="more"
-                                    aria-controls={open ? "long-menu" : undefined}
-                                    aria-expanded={open ? "true" : undefined}
-                                    aria-haspopup="true"
-                                    >
-                                    <MoreVertIcon />
-                                    </IconButton>
-                                    <Menu
-                                    id="long-menu"
-                                    MenuListProps={{
-                                    "aria-labelledby": "long-button",
-                                    }}
-                                    anchorEl={document.getElementById("long-button-menu")}
-                                    open={devGr.getOpen_menu()}
-                                    onClose={() => {
-                                    devGr.setOpen_menu(false);
-                                    }}
-                                    >
-
-                                    <MenuItem onClick={() => this.editDevice('1')}>
-                                    <ListItemIcon>
-                                    <ModeEditRoundedIcon fontSize="small" />
-                                    </ListItemIcon>{" "}
-                                    Редактировать
-                                    </MenuItem>
-
-                                    <MenuItem onClick={() => this.editDevice('2')} >
-                                    <ListItemIcon>
-                                    <LogoutRoundedIcon fontSize="small" />
-                                    </ListItemIcon>{" "}
-                                    Переместить
-                                    </MenuItem>
-                                    </Menu>
-                                    </div>
+                                   <MenuDevs/>
                                     )}
                                     </Box>
 
@@ -247,45 +154,9 @@ export class Devs extends React.Component<IProps> {
 
                                       {APP_STORAGE.getRoleWrite() === 2 &&
                                       APP_STORAGE.getRoleRead() === 1 && (
-                                      <div>
-                                      <IconButton
-                                      onClick={() => {
-                                        devGr.setOpen_menu(true);
-                                      }}
-                                      id="long-button555"
-                                      aria-label="more"
-                                      aria-controls={open ? "long-menu" : undefined}
-                                      aria-expanded={open ? "true" : undefined}
-                                      aria-haspopup="true"
-                                      >
-                                      <MoreVertIcon />
-                                      </IconButton>
-                                      <Menu
-                                      id="long-menu"
-                                      MenuListProps={{
-                                      "aria-labelledby": "long-button",
-                                      }}
-                                      anchorEl={document.getElementById("long-button555")}
-                                      open={devGr.getOpen_menu()}
-                                      onClose={() => {
-                                      devGr.setOpen_menu(false);
-                                      }}
-                                      >
-                                      <MenuItem onClick={() => this.editDevice('1')}>
-                                      <ListItemIcon>
-                                      <ModeEditRoundedIcon fontSize="small" />
-                                      </ListItemIcon>{" "}
-                                      Редактировать
-                                      </MenuItem>
 
-                                      <MenuItem onClick={() => this.editDevice('2')} >
-                                      <ListItemIcon>
-                                      <LogoutRoundedIcon fontSize="small" />
-                                      </ListItemIcon>{" "}
-                                      Переместить
-                                      </MenuItem>
-                                      </Menu>
-                                      </div>
+                                      <MenuDevs/>
+
                                       )}
                                         </Box>
                                         <Box>
@@ -307,7 +178,7 @@ export class Devs extends React.Component<IProps> {
 
                                      
                                       
-                                      <Link sx= {{fontSize:'1rem', fontWeight: '700'}} onClick={() => {window.open('http://127.0.0.1:3040/show-map?lng=' + gr.longitude + '&lat=' + gr.latitude)}}>Показать на карте
+                                      <Link sx= {{fontSize:'1rem', fontWeight: '700'}} onClick={() => {window.open('http://127.0.0.1:3040/show-map?lng=' + gr_devs[key].longitude + '&lat=' + gr_devs[key].latitude)}}>Показать на карте
                                       
                                       <IconButton color="primary" sx={{ p: '10px' }} aria-label="directions">
                                       <DirectionsIcon />
@@ -316,9 +187,6 @@ export class Devs extends React.Component<IProps> {
                                       </Box>
 
                                       </Box>
-
-  {/*------------------------- Информация об устройстве ------------------------------------------------------------*/}
-                                     
 
                                       </Box>                                   
                                       </React.Fragment>
@@ -343,13 +211,13 @@ export class Devs extends React.Component<IProps> {
 
   render(): React.ReactNode {
     return (
-      <React.Fragment>
+      <>
          <Typography sx={{ fontWeight: "500" , color: '#111111', mb : '8px'}}>
               {" "}
               Данные по устройству{" "}
          </Typography>
        {this.drawDevsFunction()}
-      </React.Fragment>
+       </>
     );
   }
 }
