+++ source/client/src/components/AppForm/Sensors/Sensors.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Box, Alert, Typography, TextField } from "@mui/material";
+import { Box, Typography} from "@mui/material";
 
 import SensorsIcon from "@mui/icons-material/Sensors";
 
@@ -14,7 +14,6 @@ import TableRow from "@mui/material/TableRow";
 
 import ArrowDownwardIcon from "@mui/icons-material/ArrowDownward";
 import ArrowUpwardIcon from "@mui/icons-material/ArrowUpward";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 
 import { TDevsGroup } from "../../../storage/components/Devs/DevEntityes";
