+++ source/show-map/src/components/show-map.tsx
@@ -1,12 +1,17 @@
+
 import React from 'react';
 import { observer } from 'mobx-react';
 
 
-import { MapContainer, TileLayer, Marker} from "react-leaflet";
+import { MapContainer, TileLayer, Marker, Popup } from "react-leaflet";
+import { Icon } from "leaflet";
 
 
-
-
+const myIcon = new Icon({
+    
+  iconUrl: './Vector4.svg',
+  iconSize: [25, 25]
+});
 const styles = {
   mapRoot: {
     height: '100vh'
@@ -36,5 +41,4 @@ export default function LeafletMap() {
     </MapContainer>
   );
 }
-
